requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
weaviate-client==4.6.3
chromadb==1.0.15
langgraph>=0.2.34
langchain>=0.3.0
langchain-openai>=0.1.17
langchain-huggingface==0.1.0
numpy==1.26.4
scikit-learn==1.5.1
annoy==1.17.3
PyYAML==6.0.2
python-json-logger==2.0.7
httpx>=0.27.0
redis==5.0.8
psycopg[binary]==3.2.1
sentence-transformers==3.0.1
tiktoken==0.7.0
spacy>=3.7.0
pytest==8.3.2
pytest-asyncio==0.23.8
PyPDF2==3.0.1
pdfplumber==0.10.3
streamlit==1.39.0
plotly==5.24.1
tqdm==4.66.1
python-dotenv==1.0.0